removed 2 unused imports took method from forms to views (the only place where it was being used)
